requirements.txt
allure-python-commons==2.13.5
attrs==23.2.0
charset-normalizer==3.3.2
colorama==0.4.6
Faker==25.8.0
idna==3.7
iniconfig==2.0.0
packaging==24.1
pluggy==1.5.0
pytest==8.2.2
python-dateutil==2.9.0.post0
requests==2.32.3
six==1.16.0